Beeper user notification counts (#50)
* Implement notification count table insert and aggregate

* Add Beeper notification count table migration

* Appease sqlite

* Fix aggregation query

* Add experimental flag to enable user notification counts

* More explicit column name

* Offset recent notification counts by 100k most recent events

Super arbitrary but this only exists as a temporary change, idea here
is that the very latest counts are much more likely to be cleared by
live receipts and we want to minimize serialization retries on the
receipt transaction.

* Don't allow local users to send receipts for unknown events

This isn't firing in production but now must be checked to prevent issues
with notification counts.

* Clear notification counts on receipt

* Fix starting value insert in migration

* Use real rooms/events in receipt tests
# Conflicts:
#	synapse/config/experimental.py